requirements.txt
Django==3.2.10
djangorestframework==3.12.4
pytest==8.2.2
pytest-django==4.8.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1